uppsrc: Fixed more Win64 warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@5822 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/HeapImp.h
@@ -1,6 +1,6 @@
 void  OutOfMemoryPanic(size_t size);
 
-void *SysAllocRaw(size_t size, int reqsize);
+void *SysAllocRaw(size_t size, size_t reqsize);
 void  SysFreeRaw(void *ptr, size_t size);
 
 void *AllocRaw4KB(int reqsize);
